backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
pymupdf
openai
firebase-admin
httpx
pytest
